:recycle: refactor: correct import from auth
